What I am happy with
This is what I'm happy with. It's simple, and I feel that it's who I am.
Over time, I will more then likely make changes to it as I learn more
about design and what I think would be a good presentation of what I
know how to do and what I think is suitable, but this is me right now.